task 16

+++ 16-binary_tree_is_perfect.c
@@ -30,7 +30,7 @@ size_t binary_tree_depth(const binary_tree_t *tree)
  * @level: Current level of the node
  * Return: 1 if tree is perfect, 0 otherwise
  */
-int is_perfect_recursive(const binary_tree_t *tree, size_t depth, size_t level)
+int is_perfect_recursive(const binary_tree_t *tree, int depth, int level)
 {
 	if (tree == NULL)
 		return (1);
@@ -53,7 +53,7 @@ int is_perfect_recursive(const binary_tree_t *tree, size_t depth, size_t level)
  */
 int binary_tree_is_perfect(const binary_tree_t *tree)
 {
-	size_t depth;
+	int depth;
 
 	if (tree == NULL)
 	{
